mxc_v4l2_capture: fix power off when closing device
There is already a call to power_off_camera() in mxc_v4l_close(),
so delete the call to vidioc_int_s_power(). This fixes a situation
where the power is off but the driver thinks it is on still.

This should be squashed with and separated from
commit: add ar1021 touchscreen

Signed-off-by: Philip Craig <phil@blackmoth.com.au>
Signed-off-by: Troy Kisky <troy.kisky@boundarydevices.com>