imx-tlv320aic3x: add support for output-enable gpio and delay
TLV320AIC3X codecs have an internal DC offset when playing the sound.
If the headphone out is connected to an amplifier, this offset will
generate a pop noise both when starting and stopping the play.
If the amplifier has an output enable pin connected to the SOM,
this feature allows to correctly manage the pin to remove that pop.
In the device tree of the board, just declare the pin and the delay:
/ {
	model = "Variscite i.MX6 DL/Solo SOM-SOLO SOM-DUAL SoloCustomBoard";
	compatible = "fsl,imx6dl-var-som", "fsl,imx6dl";

	sound {
		output-enable-gpios = <&gpio4 9 0>;
		output-enable-delay = <20>;
	};
};